requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
datasets>=2.4.0
accelerate>=0.20.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
tqdm>=4.60.0
scikit-learn>=1.0.0
